에너미+열매

+++ Assets/Atara/Enemy/Status.cs
@@ -7,5 +7,5 @@ public class Status : MonoBehaviour
     public int maxHP;
     public int currentHP;
     public float attackSpeed;
-    public float attackDamage;
+    public int attackDamage;
 }
